DELTA25/requirements.txt
flask==2.3.3
opencv-python==4.8.1.78
numpy==1.24.3
Pillow==10.0.0
boto3==1.28.40
google-generativeai==0.3.1
werkzeug==2.3.7
